chore: extract zoom functionality
