+++ src/feature/area-chart/chart.ts
@@ -12,13 +12,12 @@ import { addHours } from "date-fns";
 import EventEmitter from "eventemitter3";
 
 import { ScaleLinear, ScaleTime } from "../../types";
+import { Zoom, ZoomTransform } from "../../util/zoom";
 import { Data } from "./area-chart";
 import { AXIS_HEIGHT, AXIS_WIDTH } from "./constants";
 import { Contents } from "./contents";
 import { Colors } from "./helpers";
 import { UI } from "./ui";
-import { ZoomTransform } from "./zoom/transform";
-import { Zoom } from "./zoom/zoom";
 
 export type SeriesData = (Series<{ [key: string]: number }, string> & {
   i: (number | Date)[];
