+++ src/feature/area-chart/display-objects/vertical-axis.ts
@@ -6,8 +6,8 @@ import { Text } from "../../../renderer/text";
 import { ScaleLinear } from "../../../types";
 import { Colors } from "../../../components/depth-chart/helpers";
 import { AXIS_HEIGHT, AXIS_WIDTH, FONT_SIZE } from "../constants";
-import { Gesture } from "../zoom/gesture";
-import { Zoom } from "../zoom/zoom";
+import { Gesture } from "../../../util/zoom/gesture";
+import { Zoom } from "../../../util/zoom/zoom";
 
 export function pointer(event: any, resolution: number = 1): [number, number] {
   const node = event.target as HTMLElement;
